Split interfaces and state into folders
