+++ Compression.BSA/State/FileStateObject.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using Wabbajack.Common;
+
+namespace Compression.BSA
+{
+    public abstract class FileStateObject
+    {
+        public int Index { get; set; }
+        public RelativePath Path { get; set; }
+    }
+}
